backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.6.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
httpx==0.25.2 